requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
django-email-users==1.1.0
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
